+++ packages/components/elements/skeleton/Skeleton.tsx
@@ -20,13 +20,13 @@ function Skeleton({
     none: "hawa-rounded hawa-bg-muted",
     pulse: "hawa-animate-pulse hawa-rounded hawa-bg-muted",
     shimmer:
-      "hawa-space-y-5 hawa-rounded hawa-bg-muted  hawa-p-4 hawa-relative  before:hawa-absolute before:hawa-inset-0 before:hawa--translate-x-full before:hawa-animate-[shimmer_2s_infinite] before:hawa-bg-gradient-to-r before:hawa-from-transparent before:hawa-via-gray-300/40 dark:before:hawa-via-white/10 before:hawa-to-transparent hawa-isolate hawa-overflow-hidden  before:hawa-border-t before:hawa-border-rose-100/10"
+      "hawa-space-y-5 hawa-rounded hawa-bg-muted  hawa-p-4 hawa-relative  before:hawa-absolute before:hawa-inset-0 before:hawa--translate-x-full before:hawa-animate-[shimmer_2s_infinite] before:hawa-bg-gradient-to-r before:hawa-from-transparent before:hawa-via-gray-300/40 dark:before:hawa-via-white/10 before:hawa-to-transparent hawa-isolate hawa-overflow-hidden  before:hawa-border-t before:hawa-border-rose-100/10",
   };
   const fadeStyle = {
     bottom: "hawa-mask-fade-bottom",
     top: "hawa-mask-fade-top",
     right: "hawa-mask-fade-right",
-    left: "hawa-mask-fade-left "
+    left: "hawa-mask-fade-left ",
   };
 
   return (
@@ -36,7 +36,7 @@ function Skeleton({
         content &&
           "hawa-flex hawa-flex-col hawa-items-center hawa-justify-center",
         fade && fadeStyle[fade],
-        className
+        className,
       )}
       {...props}
     >